part3/requirements.txt
Flask==3.0.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
bcrypt==4.1.1
flask-bcrypt==1.0.1
requests==2.31.0
pytest==7.4.3
